Update Using 'ggmap' Ploting Coordinates
This data comes from a Kaggle Competition which contains the data of traffic stop time for four major cities of United States: Atlanta, Boston, Chicago and Philadelphia. I am using the ggmap and ggplot package to match all the unique coordinates in the dataset. The ggmap can be difficult to use for starters since Google started to charge the ggmap service and everybody need to register to obtain an API key and activate the Maps Static API service on https://cloud.google.com